Moved PDF APIs from ElectronicGraderController to it's own file
