+++ site/app/controllers/grading/ElectronicGraderController.php
@@ -97,15 +97,6 @@ class ElectronicGraderController extends GradingController {
             case 'remove_empty':
                 $this->ajaxRemoveEmpty();
                 break;
-            case 'pdf_annotation_fullpage':
-                $this->showPDFAnnotationFullPage();
-                break;
-            case 'pdf_annotation_embedded':
-                $this->showEmbeddedPDFAnnotation();
-                break;
-            case 'save_pdf_annotation':
-                $this->savePDFAnnotation();
-                break;
             case '':
                 $this->showStatus();
                 break;
@@ -140,71 +131,6 @@ class ElectronicGraderController extends GradingController {
             return false;
         }
         return true;
-    }
-    
-    public function savePDFAnnotation(){
-        //Save the annotation layer to a folder.
-        $annotation_layer = $_POST['annotation_layer'];
-        $annotation_info = $_POST['GENERAL_INFORMATION'];
-        $grader_id = $this->core->getUser()->getId();
-        $course_path = $this->core->getConfig()->getCoursePath();
-        $active_version = $this->core->getQueries()->getGradeable($annotation_info['gradeable_id'], $annotation_info['user_id'])->getActiveVersion();
-        $annotation_gradeable_path = FileUtils::joinPaths($course_path, 'annotations', $annotation_info['gradeable_id']);
-        if(!FileUtils::createDir($annotation_gradeable_path) && !is_dir($annotation_gradeable_path)){
-            $this->core->addErrorMessage("Creating annotation gradeable folder failed");
-            return false;
-        }
-        $annotation_user_path = FileUtils::joinPaths($annotation_gradeable_path, $annotation_info['user_id']);
-        if(!FileUtils::createDir($annotation_user_path) && !is_dir($annotation_user_path)){
-            $this->core->addErrorMessage("Creating annotation user folder failed");
-            return false;
-        }
-        $annotation_version_path = FileUtils::joinPaths($annotation_user_path, $active_version);
-        if(!FileUtils::createDir($annotation_version_path) && !is_dir($annotation_version_path)){
-            $this->core->addErrorMessage("Creating annotation version folder failed");
-            return false;
-        }
-        $new_file_name = preg_replace('/\\.[^.\\s]{3,4}$/', '', $annotation_info['file_name']) . "_" .$grader_id .'.json';
-        file_put_contents(FileUtils::joinPaths($annotation_version_path, $new_file_name), $annotation_layer);
-        return true;
-    }
-
-    public function showEmbeddedPDFAnnotation(){
-        //This is the embedded pdf annotator that we built.
-        $gradeable_id = $_POST['gradeable_id'] ?? NULL;
-        $user_id = $_POST['user_id'] ?? NULL;
-        $filename = $_POST['filename'] ?? NULL;
-        $active_version = $this->core->getQueries()->getGradeable($gradeable_id, $user_id)->getActiveVersion();
-        $annotation_path = FileUtils::joinPaths($this->core->getConfig()->getCoursePath(), 'annotations', $gradeable_id, $user_id, $active_version);
-        $annotation_jsons = [];
-        //Dir iterator needs the first file.
-        if(is_dir($annotation_path) && count(scandir($annotation_path)) > 2){
-            $first_file = scandir($annotation_path)[2];
-            $annotation_path = FileUtils::joinPaths($annotation_path, $first_file);
-            if(is_file($annotation_path)) {
-                $dir_iter = new \DirectoryIterator(dirname($annotation_path . '/'));
-                foreach ($dir_iter as $fileinfo) {
-                    if (!$fileinfo->isDot()) {
-                        $no_extension = preg_replace('/\\.[^.\\s]{3,4}$/', '', $fileinfo->getFilename());
-                        $pdf_info = explode('_', $no_extension);
-                        $pdf_id = $pdf_info[0];
-                        $grader_id = $pdf_info[1];
-                        if($pdf_id.'.pdf' === $filename){
-                            $annotation_jsons[$grader_id] = file_get_contents($fileinfo->getPathname());
-                        }
-                    }
-                }
-            }
-        }
-        $this->core->getOutput()->renderOutput(array('PDF'), 'showPDFEmbedded', $gradeable_id, $user_id, $filename, $annotation_jsons, false);
-    }
-
-    public function showPDFAnnotationFullPage(){
-        //This shows the pdf-annotate.js library's default pdf annotator. It might be useful in the future to have
-        //a full-sized annotator, so keeping this in for now.
-        $this->core->getOutput()->useFooter(false);
-        $this->core->getOutput()->useHeader(false);
-        $this->core->getOutput()->renderOutput(array('grading', 'PDFAnnotation'), 'showAnnotationPage');
     }
 
     /**
