+++ site/public/index.php
@@ -271,6 +271,10 @@ switch($_REQUEST['component']) {
         $control = new app\controllers\forum\ForumController($core);
         $control->run();
         break;
+    case 'pdf':
+        $control = new app\controllers\pdf\PDFController($core);
+        $control->run();
+        break;
     default:
         $control = new app\controllers\AuthenticationController($core, $logged_in);
         $control->run();
